+++ src/ErrorEvents.ts
@@ -1,0 +1,5 @@
+import { ErrorWithMessage } from "./ErrorWithMessage";
+
+export interface ErrorEvents<TMessage> {
+  error: (err: ErrorWithMessage<TMessage>) => unknown;
+}