+++ src/ErrorWithMessage.ts
@@ -1,0 +1,1 @@
+export type ErrorWithMessage<TMessage> = Error & { mq2?: TMessage };