+++ src/index.ts
@@ -3,4 +3,4 @@ export * from "./Consumer";
 export * from "./Publisher";
 export * from "./Subscriber";
 export * from "./Subscription";
-export * from "./types";
+export * from "./WithOptionalObjectId";