+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## NEXT
+
+- Add type-safety to all event emitters
+- Add `Consumer.drain` method (waits until consumer could not receive a message)
+
 ## 0.2.1
 
 - Fix bug in `Consumer` where past messages were sometimes not correctly consumed